Refactor some tests to use withServiceReadyOrFail util
This patch refactors to:
- move `withServiceReadyOrFail()` to helper.go
- use `withServiceReadyOrFail()` in service_to_servig test
